solution/requirements.txt
Django==1.5.4
Pillow==2.2.1
South==0.8.2
aldryn-blog==0.1.4
-e git+https://github.com/nephila/aldryn-installer@aab1af03b3b26fe6869a1d122fc4f44321dbef22#egg=aldryn_installer-master
dj-database-url==0.2.2
django-classy-tags==0.4
git+git://github.com/divio/django-cms.git#egg=django-cms
django-mptt==0.5.2
django-reversion==1.7.1
django-sekizai==0.7
djangocms-admin-style==0.1.4
djangocms-column==1.1.2
djangocms-style==1.0.6
djangocms-text-ckeditor==2.0.2
html5lib==0.99
six==1.4.1
wsgiref==0.1.2
